[Feature] ScoreCount Long으로 변경

+++ src/main/java/com/example/backend/lecture/entity/dto/response/LectureListResponseDTO.java
@@ -31,7 +31,7 @@ public class LectureListResponseDTO {
 
     private double averageScore;
 
-    private int scoreCount;
+    private Long scoreCount;
 
     private List<LectureImage> imageUrl;
 }
